Move paCallbackData and TxRxThread to pipeline library.
